refactor: 파일명 및 구조 변경
